Jhwang2

+++ srcs/block/server.cpp
@@ -15,10 +15,10 @@ Server& Server::operator =(const Server& src) {
 	return (*this);
 }
 
-void						Server::setPort(uint32_t _port) { port = _port; }
+void						Server::setPort(size_t _port) { port = _port; }
 void						Server::setDomain(std::string _domain){ domain = _domain; }
 
-uint32_t					Server::getPort(void) const { return port; }
+size_t					Server::getPort(void) const { return port; }
 const std::string&			Server::getDomain(void) const{ return domain; }
 std::list<Location>&		Server::getLocationList(void) { return location_list; }
 const std::list<Location>&	Server::getLocationListConst(void) const { return location_list; }
